+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - A tag replacement for Structure Nav to better support new template languages and GraphQL
 - A method for embedding native ExpressionEngine templates within Twig and Blade
 - Field content implements countable for better support with Twig's length filter
+- Page variables to Channel Entries Tag results
 
 ### Fixed